[Refactor]: City 위도,경도 float 타입으로 변경

+++ src/main/java/com/minair/domain/City.java
@@ -22,6 +22,6 @@ public class City {
     private Long id;
 
     private String name;
-    private double longitude;
-    private double latitude;
+    private float longitude;
+    private float latitude;
 }
